Rearrange some tests
